Updated genesis block for testnet
Pushing files in "test net" mode (official block is set to a test net genesis block) so that we can test the hard-fork on the testnet before migrating to the real network. In this form the client will not connect to the "main" block chain, only the test net chain.